combine modal and form components
